Update deps 2024-01 (#1225)
* Add a helper script for updating dependencies

* Update dependencies for repo root

* Update dependencies for system-tests

* Update dependencies for cms

* Update dependencies for course-material

* Update dependencies for example-exercise

* Update dependencies for main-frontend

* Update dependencies for quizzes

* Update dependencies for tmc

* Update dependencies for shared-module

* npm audit for shared-module

* Cargo upgrade

* Cargo update

* Update dependency update script

* Cargo check

* Clippy fixes

* Fix snippet annotations being disabled on the newer versions of ingress-nginx

* Try removing the date interceptor

* Refactor ui code to work with the new dates

* Updated ts-rs

* Improve bin help messages

* Generate doc files

* Fix Gutenberg type generation

* Type fixes

* Render stats in /stats

* Add a link to the new stats page

* Downgrade next in course material because of a miscompilation

* System test fixes

* Eslint fixes

* Generate doc files

* Fix

---------

+++ services/headless-lms/utils/src/cache.rs
@@ -34,7 +34,7 @@ impl Cache {
                     return false;
                 };
                 match conn
-                    .set_ex::<_, _, ()>(key, value, expires_in.as_secs() as usize)
+                    .set_ex::<_, _, ()>(key, value, expires_in.as_secs())
                     .await
                 {
                     Ok(_) => true,
